Document APIs in ByteString/Text modules
I'm doing this because I landed on the package in hackage, went to the
module I needed (Text), didn't understand one of the arguments, read the
source, then had to go to common to find what the argument was. I
figured it would be nice if all the modules the user was likely to use
were documented equally.